move username fetch back into parseFrontMatter
